docs: one boolean per strategy
