--- libs/iresearch/include/iresearch/search/docs/make_bitset_conjunction.cpp
@@ -1,35 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-/// DISCLAIMER
-///
-/// Copyright 2026 SereneDB GmbH, Berlin, Germany
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     http://www.apache.org/licenses/LICENSE-2.0
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-/// Copyright holder is SereneDB GmbH, Berlin, Germany
-////////////////////////////////////////////////////////////////////////////////
-
-#include "iresearch/index/index_reader.hpp"
-#include "iresearch/search/common/conjunction_bitset.hpp"
-#include "iresearch/search/docs/bitset.hpp"
-#include "iresearch/search/docs/plan.hpp"
-
-namespace irs::docs {
-
-Root::ptr MakeBitsetConjunction(std::span<const search::PostingClause> terms,
-                                std::span<const QueryBuilder::ptr> filters,
-                                const SubReader& segment, const Context& ctx) {
-  return search::MakeConjunctionBitset<Root::ptr>(terms, filters, nullptr,
-                                                  segment, nullptr);
-}
-
-}  // namespace irs::docs
